Add `sacompose` type hints
Technically this approach has an edge case where it would
get the type hints wrong (if one function in `sacompose`
returns `Awaitable[T]`, and the next function in the
composition takes `Awaitable[T]` as its first parameter).

But this is the best we can do with the type hinting that
Python currently provides, short of a combinatoric
explosion of overloads (we'd need an extra `-> Never`
overload for every argument of every arity).